Conflicts:
	exponent_version.php
	external/editors/ckeditor/plugins/sourcedialog/dialogs/sourcedialog.js
	external/editors/ckeditor/plugins/sourcedialog/icons/sourcedialog-rtl.png
	external/editors/ckeditor/plugins/sourcedialog/lang/af.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ar.js
	external/editors/ckeditor/plugins/sourcedialog/lang/bg.js
	external/editors/ckeditor/plugins/sourcedialog/lang/bn.js
	external/editors/ckeditor/plugins/sourcedialog/lang/bs.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ca.js
	external/editors/ckeditor/plugins/sourcedialog/lang/cs.js
	external/editors/ckeditor/plugins/sourcedialog/lang/cy.js
	external/editors/ckeditor/plugins/sourcedialog/lang/da.js
	external/editors/ckeditor/plugins/sourcedialog/lang/de.js
	external/editors/ckeditor/plugins/sourcedialog/lang/el.js
	external/editors/ckeditor/plugins/sourcedialog/lang/en-au.js
	external/editors/ckeditor/plugins/sourcedialog/lang/en-ca.js
	external/editors/ckeditor/plugins/sourcedialog/lang/en-gb.js
	external/editors/ckeditor/plugins/sourcedialog/lang/en.js
	external/editors/ckeditor/plugins/sourcedialog/lang/eo.js
	external/editors/ckeditor/plugins/sourcedialog/lang/es.js
	external/editors/ckeditor/plugins/sourcedialog/lang/et.js
	external/editors/ckeditor/plugins/sourcedialog/lang/eu.js
	external/editors/ckeditor/plugins/sourcedialog/lang/fa.js
	external/editors/ckeditor/plugins/sourcedialog/lang/fi.js
	external/editors/ckeditor/plugins/sourcedialog/lang/fo.js
	external/editors/ckeditor/plugins/sourcedialog/lang/fr-ca.js
	external/editors/ckeditor/plugins/sourcedialog/lang/fr.js
	external/editors/ckeditor/plugins/sourcedialog/lang/gl.js
	external/editors/ckeditor/plugins/sourcedialog/lang/gu.js
	external/editors/ckeditor/plugins/sourcedialog/lang/he.js
	external/editors/ckeditor/plugins/sourcedialog/lang/hi.js
	external/editors/ckeditor/plugins/sourcedialog/lang/hr.js
	external/editors/ckeditor/plugins/sourcedialog/lang/hu.js
	external/editors/ckeditor/plugins/sourcedialog/lang/is.js
	external/editors/ckeditor/plugins/sourcedialog/lang/it.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ja.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ka.js
	external/editors/ckeditor/plugins/sourcedialog/lang/km.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ko.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ku.js
	external/editors/ckeditor/plugins/sourcedialog/lang/lt.js
	external/editors/ckeditor/plugins/sourcedialog/lang/lv.js
	external/editors/ckeditor/plugins/sourcedialog/lang/mn.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ms.js
	external/editors/ckeditor/plugins/sourcedialog/lang/nb.js
	external/editors/ckeditor/plugins/sourcedialog/lang/nl.js
	external/editors/ckeditor/plugins/sourcedialog/lang/no.js
	external/editors/ckeditor/plugins/sourcedialog/lang/pl.js
	external/editors/ckeditor/plugins/sourcedialog/lang/pt-br.js
	external/editors/ckeditor/plugins/sourcedialog/lang/pt.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ro.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ru.js
	external/editors/ckeditor/plugins/sourcedialog/lang/sk.js
	external/editors/ckeditor/plugins/sourcedialog/lang/sl.js
	external/editors/ckeditor/plugins/sourcedialog/lang/sq.js
	external/editors/ckeditor/plugins/sourcedialog/lang/sr-latn.js
	external/editors/ckeditor/plugins/sourcedialog/lang/sr.js
	external/editors/ckeditor/plugins/sourcedialog/lang/sv.js
	external/editors/ckeditor/plugins/sourcedialog/lang/th.js
	external/editors/ckeditor/plugins/sourcedialog/lang/tr.js
	external/editors/ckeditor/plugins/sourcedialog/lang/ug.js
	external/editors/ckeditor/plugins/sourcedialog/lang/uk.js
	external/editors/ckeditor/plugins/sourcedialog/lang/vi.js
	external/editors/ckeditor/plugins/sourcedialog/lang/zh-cn.js
	external/editors/ckeditor/plugins/sourcedialog/lang/zh.js
	external/editors/ckeditor/plugins/sourcedialog/plugin.js
	external/editors/ckeditor/plugins/sourcedialog/samples/sourcedialog.html
	framework/core/subsystems/expCSS.php
	framework/core/subsystems/expTheme.php
	framework/modules/ecommerce/views/eventregistration/show.tpl
	framework/modules/events/controllers/eventController.php
	framework/modules/forms/controllers/formsController.php
	framework/modules/motd/views/motd/show.tpl
	framework/modules/navigation/controllers/navigationController.php
	framework/modules/twitter/controllers/twitterController.php
	framework/modules/twitter/models/Twitter.php
	framework/modules/twitter/models/expTwitter.php
	framework/plugins/function.help.php
	install/upgrades/upgrade_mediaplayer.php

+++ framework/modules/file/models/expFile.php
@@ -401,7 +401,7 @@ class expFile extends expRecord {
 
         // If $_destFile is defined, use that name as an override for the
         // uploaded file name
-        $_destFile = ($_destFile == null) ? self::fixFileName($_FILES[$_postName]['name']) : $_destFile;
+        $_destFile = ($_destFile == null) ? self::fixName($_FILES[$_postName]['name']) : $_destFile;
 
         // Fix the filename, so that we don't have funky characters screwing
         // with our attempt to create the destination file.
@@ -518,7 +518,7 @@ class expFile extends expRecord {
 
         // If $_destFile is defined, use that name as an override for the
         // uploaded file name
-        $_destFile = ($_destFile == null) ? self::fixFileName($fileName) : $_destFile;
+        $_destFile = ($_destFile == null) ? self::fixName($fileName) : $_destFile;
 
         // Fix the filename, so that we don't have funky characters screwing
         // with our attempt to create the destination file.
@@ -727,9 +727,99 @@ class expFile extends expRecord {
      * @throws void
      *
      */
-    public static function fixFileName($name) {
-//        return preg_replace('/[^A-Za-z0-9\.]/','_',$name);
-        return preg_replace('/[^A-Za-z0-9\.]/', '-', $name);
+    public static function fixName($name) {
+        return preg_replace('/[^A-Za-z0-9\.]/','_',$name);
+//        return preg_replace('/[^A-Za-z0-9\.]/', '-', $name);
+    }
+
+    /**
+     * Return the mimetype for the passed filename
+     *
+     * @param string $filename
+     * @return string
+     */
+    public static function getMimeType($filename) {
+        /* Store an array of commom mimetypes */
+        $types = array(
+        'txt' => 'text/plain',
+        'htm' => 'text/html',
+        'html' => 'text/html',
+        'php' => 'text/html',
+        'css' => 'text/css',
+        'js' => 'application/javascript',
+        'json' => 'application/json',
+        'xml' => 'application/xml',
+
+        // images
+        'png' => 'image/png',
+        'jpe' => 'image/jpeg',
+        'jpeg' => 'image/jpeg',
+        'jpg' => 'image/jpeg',
+        'gif' => 'image/gif',
+        'bmp' => 'image/bmp',
+        'ico' => 'image/vnd.microsoft.icon',
+        'tiff' => 'image/tiff',
+        'tif' => 'image/tiff',
+        'svg' => 'image/svg+xml',
+        'svgz' => 'image/svg+xml',
+
+        // archives
+        'gz' => 'application/x-gzip',
+        'zip' => 'application/zip',
+        'rar' => 'application/x-rar-compressed',
+        'exe' => 'application/x-msdownload',
+        'msi' => 'application/x-msdownload',
+        'cab' => 'application/vnd.ms-cab-compressed',
+
+        // audio/video
+        'mp3' => 'audio/mpeg',
+        'ogg' => 'audio/ogg',
+        'qt' => 'video/quicktime',
+        'mov' => 'video/quicktime',
+        'f4v' => 'video/mp4',
+        'mp4' => 'video/mp4',
+        'm4v' => 'video/x-m4v',
+        'ogv' => 'video/ogg',
+        '3gp' => 'video/3gpp',
+        'webm' => 'video/webm',
+        'flv' => 'video/x-flv',
+        'swf' => 'application/x-shockwave-flash',
+
+        // adobe
+        'pdf' => 'application/pdf',
+//        'psd' => 'image/vnd.adobe.photoshop',
+//        'ai' => 'application/postscript',
+//        'eps' => 'application/postscript',
+//        'ps' => 'application/postscript',
+
+        // ms office
+//        'doc' => 'application/msword',
+//        'rtf' => 'application/rtf',
+//        'xls' => 'application/vnd.ms-excel',
+//        'ppt' => 'application/vnd.ms-powerpoint',
+
+        // open office
+//        'odt' => 'application/vnd.oasis.opendocument.text',
+//        'ods' => 'application/vnd.oasis.opendocument.spreadsheet'
+        );
+
+        /* Get the file extension,
+         * FYI: this is *really* hax.
+         */
+        $extension = strtolower(array_pop(explode('.',$filename)));
+        if(array_key_exists($extension, $types)) {
+            /* If we can *guess* the mimetype based on the filename, do that for standardization */
+            return $types[$extension];
+        } elseif(function_exists('finfo_open')) {
+            /* If we don't have to guess, do it the right way */
+            $finfo = finfo_open(FILEINFO_MIME);
+            $mimetype = finfo_file($finfo, $filename);
+            finfo_close($finfo);
+            return $mimetype;
+        } else {
+            /* Otherwise, let the browser guess */
+            return 'application/octet-stream';
+        }
     }
 
     /**
@@ -1485,10 +1575,6 @@ class expFile extends expRecord {
         rmdir($dir);
     }
 
-    public static function fixName($name) {
-        return preg_replace('/[^A-Za-z0-9\.]/', '_', $name);
-    }
-
     /** exdoc
      * Move an uploaded temporary file to a more permanent home inside of the Exponent files/ directory.
      * This function takes into account the default file modes specified in the site configuration.
@@ -1652,12 +1738,12 @@ class expFile extends expRecord {
      * @return string
      * @node Model:expFile
      */
-    public static function dumpDatabase($db, $tables = null, $force_version = null) {
+    public static function dumpDatabase($db, $tables = null, $type = null, $record = null) {
         $dump = EQL_HEADER . "\r\n";
-        if ($force_version == null) {
+        if ($type == null) {
             $dump .= 'VERSION:' . EXPONENT . "\r\n\r\n";
         } else {
-            $dump .= 'VERSION:' . $force_version . "\r\n\r\n";
+            $dump .= 'VERSION:' . EXPONENT . ':' . $type . "\r\n\r\n";
         }
 
         if (!is_array($tables)) {  // dump all the tables
@@ -1676,7 +1762,15 @@ class expFile extends expRecord {
             $tabledef = $db->getDataDefinition($table);
             $dump .= 'TABLE:' . $table . "\r\n";
             $dump .= 'TABLEDEF:' . str_replace(array("\r", "\n"), array('\r', '\n'), serialize($tabledef)) . "\r\n";
-            foreach ($db->selectObjects($table) as $obj) {
+            $where = '1';
+            if ($type == 'Form') {
+                if ($table == 'forms') {
+                    $where = 'id=' . $record;
+                } elseif ($table == 'forms_control') {
+                    $where = 'forms_id=' . $record;
+                }
+            }
+            foreach ($db->selectObjects($table,$where) as $obj) {
                 $dump .= 'RECORD:' . str_replace(array("\r", "\n"), array('\r', '\n'), serialize($obj)) . "\r\n";
             }
             $dump .= "\r\n";
@@ -1702,7 +1796,7 @@ class expFile extends expRecord {
      * @return bool
      * @node Model:expFile
      */
-    public static function restoreDatabase($db, $file, &$errors, $force_version = null) {
+    public static function restoreDatabase($db, $file, &$errors, $type = null) {
         $errors = array();
 
         if (is_readable($file)) {
@@ -1714,13 +1808,12 @@ class expFile extends expRecord {
                 return false;
             }
 
-            if ($force_version == null) {
-                $version = explode(':', trim($lines[1]));
-                $eql_version = $version[1] + 0;
-            } else {
-                $eql_version = $force_version;
-            }
+            $version = explode(':', trim($lines[1]));
+            $eql_version = $version[1] + 0;
             $current_version = EXPONENT + 0;
+            if ((array_key_exists(2, $version) && $type == null) || $version[2] != $type) {
+                $eql_version = 0;  // trying to import wrong eql type
+            }
 
 //            $clear_function = '';
             $fprefix = '';
@@ -1760,7 +1853,9 @@ class expFile extends expRecord {
                             $table_function = $fprefix . $table;
                         }
                         if ($db->tableExists($table)) {
-                            $db->delete($table);  // drop/empty table records
+                            if ($type == null) {
+                                $db->delete($table);  // drop/empty table records
+                            }
 //                            if ($clear_function != '') {
 //                                $clear_function($db, $table);
 //                            }
@@ -1807,6 +1902,20 @@ class expFile extends expRecord {
                             $pair[1] = str_replace('\r\n', "\r\n", $pair[1]);
     //						$object = expUnserialize($pair[1]);
                             $object = @unserialize($pair[1]);
+                            if ($type == 'Form') {
+                                if ($table == 'forms') {
+                                    $forms_id = $object->id = $db->max($table,'id') + 1;  // create a new record
+                                    $spare = new expRecord();
+                                    $spare->title = $object->title;
+                                    $spare->makeSefUrl();
+                                    $object->sef_url = $spare->sef_url;
+                                } elseif ($table == 'forms_control') {
+                                    $object->id = null;  // create a new record
+                                    $object->forms_id = $forms_id;  // assign to new form record
+                                } elseif (substr($table,6) == 'forms_') {
+                                    $object->id = null;  // create a new record
+                                }
+                            }
                             if (!$object) $object = unserialize(stripslashes($pair[1]));
                             if (function_exists($table_function)) {
                                 $table_function($db, $object);
@@ -1873,29 +1982,6 @@ class expFile extends expRecord {
                 }
             }
 
-            // check for and process to rebuild new forms module data table
-            if (!empty($newformdata)) {
-                foreach ($newformdata as $tablename=>$tabledata) {
-                    $newform = $db->selectObject('forms','table_name="'.substr($tablename,6).'"');
-                    if (!empty($newform)) {
-                        // create the new table
-                        $form = new forms($newform->id);
-                        $table = $form->updateTable();
-
-                        // populate the table
-                        foreach ($tabledata as $record) {
-                            $record = str_replace('\r\n', "\r\n", $record);
-                            $object = @unserialize($record);
-                            if (!$object) $object = unserialize(stripslashes($record));
-                            if (is_object($object)) {
-                                $db->insertObject($object, 'forms_' . $table);
-                            }
-                        }
-                        $errors[] = sprintf(gt('*  However...we successfully recreated the "%s" Table from the EQL file'), $table);
-                    }
-                }
-            }
-
             // rename mixed case tables if necessary
             expDatabase::fix_table_names();
 //            if ($eql_version != $current_version) {
@@ -1917,8 +2003,8 @@ class expFile extends expRecord {
     }
 
     /**
-     * recreates an deprecated formbuilder data table
-     * needed to import form data from eql file exported prior to v2.2.0
+     * recreates a deprecated formbuilder data table
+     * needed to import form data from eql file exported prior to v2.1.4
      * this is just the old formbuilder_form::updateTable method
      *
      * @static
